installer: Add post-install and pre-install steps for rpm install
In post-install, we want to populate the node.properties file
originally templated with the node.id to have the actual value of uuidgen
on the node. This will enable us to have a single-node presto installed
by just running rpm -i. No manual configuration will be needed. Also
deleted the node.properties.template file since this will be unused now.

In pre-install, we check if Java installed is Oracle's java version 1.8.
The install will fail if this vendor and version is not found.
We did not want to add jdk as the presto rpm pre-requisite, hence this check.

Task: Swarm-58
Test: Manual test on EC2 with and without Java 8 installed.