installer: Get the configured data-dir launcher config in init.d scripts
Summary:
We were hard-coding the --data-dir config argument to bin/launcher.py
in the init.d scripts. This patch fixes the logic to get it from
/etc/presto/node.properties file (we assume that config is deployed here
always)

We also append a timestamp_pid id to the log file names. Note that the
pid is current shell pid of the init.d script (not the actual Presto
server pid)

Task: Swarm-60

Test Plan: manual

Differential Revision: https://phabricator.td.teradata.com/D132