Better languageTag generation in JDBC http client
Summary:
As JDBC driver is written in Java6 we cannot use Locale.getLanguageTag()
method which was introduced in Java7 for passing locale information in
HTTP header from clent to Presto server.

This patch provides implementation of languageTag generation. This
implementation is not IETF BCP 47 compliant but still strings generated with it
are properly desrialized to Locale objects when Locale.forLanguageTag
method is used on server side.

Issue: SWARM-551

Test Plan: Build

Reviewers: pnowojski

Differential Revision: https://phabricator.td.teradata.com/D521