Docs: Document rpm build and associated installer related changes
Summary:
We add the installer related changes to a README.txt for now.
presto-server/ already had a README.txt which is deployed on the cluster
at /usr/share/docs/presto/. So adding this README under
presto-server/resources. Later we should move/integrate it with other
docs.

SWARM-661

Test Plan: none

Differential Revision: https://phabricator.td.teradata.com/D695