Refactor SPI docs and add more function examples
Refactor SPI docs for clarity, extracting the generic SPI parts to the
overview and creating a new page for connector specific things. Add
more detailed examples to the function documentation so that it is
possible to just copy-paste from the documentation into a plugin
and have it work.

Testing: built docs