Implement DEALLOCATE PREPARE
Syntax:
DEALLOCATE PREPARE <name>

Confirms existence of prepared statement with name <name> and adds the
statement to a set of deallocated prepared statements.  Sends the set of
deallocated statements back to the client in an HTTP header.  The
client is responsible for maintaining the prepared statements for a
session.

Testing: added unit tests