"avg" aggregate function for decimal(p,s)
This is the first of the aggregate functions for parametric types.
Implemented using the "old" way, via specialize() not annotations, since
that "new" mechanism doesn't support parametric types for aggregates
yet. Probably to be refactored after improving latter.

This implementation follows the other aggregate functions implementation
patterns (like having the state object initialized in the function, etc).

The return value of the "avg" function is a decimal with the scale and
precision of the input column. The implementation uses a BigInteger
internally as the accumulator.

Conflicts:
  presto-spi/src/main/java/com/facebook/presto/spi/type/LongDecimalType.java