"round" scalar functions for decimal(p,s)
There are two variants of this method round(num) and round(num, scale).

The first is a shorthand for round(num, 0). However currently those
methods return different results in terms of output precision and scale.
Due to an internal temporary limitation (not being able to evaluate the
scale literal parameter at semantic analysis time) the result scale of
the round(num) result is 0, while the precision is fitted to store a
possible additional digit on the left. The result of round(num, scale)
is always the original input scale and precision (+1 for the possible
rounding result).

Negative scale is permitted, allows to clear the positions left of the
decimal dot with proper rounding.

Conflicts:
	presto-main/src/test/java/com/facebook/presto/operator/scalar/TestMathFunctions.java