Resolve ambiguities for the operator invocations with unknown type
As we have introduced the most specific function selection algorithm,
that throws an error in the situation when the most specific method
can't be selected, we must resolve the conflicts explicitly for the
UNKNOWN type operators.

For instance for the invokation `BIGINT + UKNNOWN` two operators are
selected as applicable:

- `ADD(BIGINT, BIGINT)`
- `ADD(BIGINT, INTERVAL)`

Before the `most specific function` algorithm was introduced the
random operators from those two was considered for invokation.

Most specific function selection algorithm can't resolve this conflict
because the `BIGINT` and the `INTERVAL` types are not linked with the
implicit coercion. For the algorithm theese two types are just 2 completely
different types. But bassicaly the default sematic for UNKNOWN should be
simple. `OPERATOR(T, UNKNOWN)` must return `T`. To keep this semantic
and resolve the ambigulity it was decided to introduce explicit
arithmetic and inequallity operators for the `UNKNOWN` type.