Add TestHiveStorageFormats product test
Test INSERT and SELECT paths for all the storage formats
supported by hive connector. This test is going to be used
for kerberized HDFS access verification.

Initally all the formats are commented out because no one of them
is supported yet. We will uncomment format by format together with
the kerberos support implmenetation pathches. So we can better track
what changes are needed in order to ensure that some particular format
works for Kerberized Hadoop.