Quarantine testSelectPartitionedHiveTableDifferentFormats
The test uses the Presto REST API to confirm the number of rows processed.
Occasionally, the number of rows read will not be reported correctly. We
believe this happens because of a race condition in the propagation of
statistics. A more thorough investigation will be done after our 141t release.