Change download link text to www.teradata.com/presto
The download links had the text of the software implying that it was a direct download link. However, it directed one to www.teradata.com/presto making it seem like the link was broken. The text has been changed to indicate that the software package can be downloaded by going to teradata.com/presto.