Resolve JSON functions ambiguities for UNKNOWN arguments
`JSON` and `VARCHAR` types are not correlated with the implicit coercion.
So for example `jsonSize(NULL)` expression is ambiguous. For such expression
either `jsonSize(VARCHAR)` or `jsonSize(JSON)` can be invoked. Similar
ambiguities are present for the `json_array_contains` and the `json_size` methods
as well. To resolve them explicit operators that accept `UNKNOWN` as a parameters
have been introduced.

Actually better approach is to remove all the `VARCHAR` versions of the `JSON`
functions and introduce implicit coercion from `VARCHAR` to `JSON` similarly
to coercion from `VARCHAR` to `JSON_PATH`. But to do that we should change the
`varchar to json cast` semantic to invoke `json_parse` for the cast instead of
just wrapping string into the quotes.