Add product tests with mergeable WindowNodes
WindowNodes that have identical specifications can be merged. This adds product
tests covering a number of situations with mergeable windows. Note that window
functions that take extra parameters result in a different plan than ones that
do not. nth_value(), for example, takes a parameter specifying which value in
the window to take.

These functions produce a plan with a ProjectNode that cannot be pruned by the
PruneIdentityProjections optimizer, and so are special cases that should be
tested.