Fix problem with `round` on short int types casts
Math functions defines multiple versions of two argument
`round` function:
 1) `round(TINYINT, BIGINT)`
 2) `round(SMALLINT, BIGINT)`
 3) `round(INTEGER, INTEGER)`
 4) `round(BIGINT, BIGINT)`

There is an incosequence in (3), as the second argument is
suprisingly INTEGER. This was only a small problem before
`TINYINT` and `SMALLINT` were introduced.

Right now it causes a problem with coersions when for
example `round(TINYINT, TINYINT)` is called.
In such scenario both (3) and (4) are valid coersions
and so such call is ambiguous (causes an error).

To fix this we may either define (generate) 16 function
for all possible combination or just make second argument
`BIGINT` for all cases (which is proposed change).

The proposed solution has a drawback of calling cast
in some cases, but this cast is noop in current implementation.
The return type with 4 functions is determined correctly only
based on first argument type.

This commit also introduces test for smaller integer type
for `round` function which do cover introduced change.