Add execution support for execute with using
Parameters get replaced when you create the execution task or plans.
For regular sql query execution, this means that at analysis time, you
pass in the list of parameters for analysis, similar to how
QualifiedNameReferences work so that they can be analyzed correctly.
The expressions get rewritten and coerced as needed whenever you add the
expression to a plan node.  For explain tasks, parameter values are also
passed through.  They are replaced in sqlFormatting and planning.  For data
definition tasks, they get replaced as part of the execute method.