Implement group based authorization for OpenLDAP
Once the user's password is validated, the user's credentials will be
used to check if the user belongs to a particular group. The server
property 'authentication.ldap.group-dn' must have the group's
distinguished name.

This commit handles group check for OpenLDAP.

'memberof' overlay is assumed to be enabled for OpenLDAP.
The docker containers running the product-tests will have that
enabled.