Unquarantine TPC-H q19
Presto for TPC-H q19 and data returned from tpch connector returns the
same result as postgresql. That makes this query working correctly,
however the result is unacceptable according to TPC-H specification.

This is could be due tpch connector generates unacceptable by TPC-H
specification data, this idea has to be verified.